Setup a minimal theme github pages (#2)
This adds a github page for the organization that can be seen from blockycraft.jrbeverly.dev. The content is a stub, but uses the blockycraft icon for a favicon.

The webpage will eventually be used to host either the Unity webplayer in some manner. Potentially as an embedded link or something, so that it can be easily distribution.