Voxel resource pack textures included from Kenney.nl (#7)
Bundle in the texture assets pulled from Kenney's voxel pack. These will be used to describe faces of each voxel/block in the world.

Instead of the full pack of textures, I have included just the block types. It happened to work out to be a NxN grid, but the code will need to support NxM grids.