Proof of concept work revised (#9)
A line of blocks rendered in a scene with a controllable camera. The block types are all of the 'grass' type, as seen in the Kenney.nl voxel pack.

Each of the block types is defined as an asset, and can have the sides and faces customized. Additional work can be done to create the idea of a texture pack.