Build and deploy to gh-pages with github actions (#15)
Compile the unity game and deploy to github pages, accessible using the Unity WebGL player. The license defined for use by CI is that of personal rather than professional.

The default scene for the project was not set, which meant that CLI-based compilation would just return a black-screen. This ensures that the block-world is shown on startup.

The compiled WebGL application is deployed to blockycraft.jrbeverly.dev, available under /player.