Treat the chunk as a 2D grid (#16)
Previous the offset was computed as a line, this switches it to a 2D grid. The completed visual is just a flat grid of grass items.

The end-aim is for this to be 3D, but 2D is a necessary intermediate step to get a grasp on iteration and interfacing for chunks.