Convert chunks from 2D plane to 3D plane (#18)
Converts the 2d voxel plane to a 3d dimensional plane. This commits supports new means of describing blocks in the world, and iterating through the chunk space.

As a note on performance: This implementation is very inefficient in terms of object creation, C# language structs (IEnumerable & multi-dimensional arrays). I'm fine with this for now, as I am trying to narrow down an API that is convenient for world manipulation. The first Blockycraft (& redux version) had significant issues because of how the world accessors behaved. I'd prefer to work in inefficient and malleable, than stiff and performant.