Chunks as children of the World instance (#21)
The Chunk object does not need to be a monobehaviour. This moves responsibilities of the Chunk into its parent World, that will allow for better chunk generation tooling.

This is starting to simplifying the chunk object so that creation can be handled by a factory (e.g. procedural generation tooling)