Move GitHub Pages deploy to separate job (#34)
Move deploying to `gh-pages` into its own job, that will execute after the build process has completed.

The build job is only responsible for deploying to `gh-pages` when running on `master` and `WebGL`. This adds overhead when working with other target platforms, as many of the jobs related to `gh-pages` or releases are not relevant.