Support a 3D Iterator model replacing block enumeration (#36)
Support an iterator capable of simplifying iterator over three dimensions into a single iteration loop.

This keeps with the cleanliness of the previous approach, while removing the unnecssary bloat that the `Block` per loop introduced. Both of these approaches create unnecessary allocations through types (previously `Block`, this by `anonymous type`). However `Iterator3D` removes locking the concept of Iterator to Blocks, and replaced it with a generic Iterator.

A `fast` iterator may be considered using a model like `.HasNext(), .Next(), .Peek()`, but in terms of performance considerations I do not believe they will be necessary for the project.