Documentation regarding the classic implementation (#37)
Configure documentation that refers to the classic version of the project. explaining the origins of the project.

The classic version is referenced by the documentation, as I wanted to provide some context as to why there are two 'versions' of the codebase. Classic will only show up in documentation, as the new direction with Unity is an improved development experience.