Simple continuous world based on dictionary model (#39)
Configure a simple continous world based on a dictionary for world create.

A dictionary was a good starting point as the code doesn't require any complex querying of the chunk system. For now, it just needs to know if a chunk already exists, which is a good fit for a dictionary. This will be shifted to a different model for rendering, as the concept of a Chunk is (at this point) offering a pretty poor interface.

Message passing between the world and player should follow the Unity model of 'Messages' (or the more threaded approach). For now, this is sufficient for experimenting with the issues an infinite world can create.