Reduce number of faces created by block visibility (#43)
Reduce the number of faces built by VoxelBuilder based on a simple visibility check.

This check doesn't need to be overly complex, and can work with local data only. The main intention is to unblock increases in the number of chunks + blocks per chunk. Moving forward the VoxelBuilder should not be knowledgeable on the "space" of Blocks, but instead rely on being passed in the blocks to render (a la Factory style).

This unblocks issues with the maximum number of blocks per chunk. We cannot increase the number of blocks to 32x32x32 without running up against the maximum amount of data that can be put in a mesh. When scaling up to this size, graphics issues occur.