Convert WorldGenerator to Asset type instead of static classic (#46)
Convert the WorldGenerator from a static-ish class to an asset type bundled with asset:biome.

This allows biomes to set the generator they wish to leverage, and have custom properties based on this. Generators can then be swapped out for each individual biome or customized as necessary to optimize the generation pipeline.

Additionally added a sand biome as a second biome just for getting an idea of how these two will work together. At the moment it randomly selected one of the biomes using a simple Random.value call. This is not intended to be long term.