Support padding and metadata for blocksheet (#47)
Support padding and metadata for the Blocksheet. The metadata is intended to replace the texture enumeration.

The enumeration uses an index-to-coordinate mapping scheme (idx -> (x,y)) that relies on all of the block textures appearing in the exact same order. If for some reason one of these blocks were to be re-ordered, it would require fixing the IDs of all values in the enumeration.

To avoid this, it makes more sense to just store the location data of each block (x=2, y=2, w=128, h=128), then compute the associated UV values from these. This removes any reliance on the ordering of objects in the spritesheet. This change will come later, as this is just a starter for cleaning up some of the assets.

The UV mapping for the blocks has not been adjusted to take into account the 2 padding on all images. This means that there will be a visual issue on all chunks until the switch-over has been completed.