Convert BlockTypeTexture to data assets (#49)
Convert the code enumeration BlockTypeTexture to the TexturePack.Element data type that can be stored as an asset.

This removes the need for a code change when making adjustments to the available textures, and removes the ID pinning requirement. The defined textures are now available as objects that can be referenced by there slug name (e.g. grass_top).

The BlockType asset now requires the name of the texture to be provided, rather than an index value. This means all of the block type assets have been updated to reflect this change.