Introduce the air block type, defining the idea of empty space (#55)
Introduce an 'Air' block type, and add this type to the sand and flat biomes for reference.

Resolve a discrepancy with how the system checked for neighbours, and how the chunk factory positioned those people. The fix ensures that the concept of a Neighbour maps correctly to the visual neighbour location.

Although having null represent 'Air' does offer advantages, those don't really match that well when invisible blocks exist. A barrier block is a good example. This is a block that has no visible properties, but still has a bounding box.