Coordinate system adjustments to better map visual & logical positions (#61)
Match the visual and logical coordinate systems to operate in same dimensions.

The previous system had the dimensions configured in an odd way. This meant as the X value increased, it would be represented in the negative dimension in the visual representation. This complicates a lot of the work for block placement/removal and determining the location of the player.