Abstract away concept of dictionary from world management (#64)
Abstract away the dictionary from world management, allowing a common API for 3D world accessors.

The world is managed using a dictionary at the moment, which works pretty well. It allows for efficient access of nearby chunks, and doesn't have additional memory requirements that a 3d array might incur. For a short-term solution for managing the available chunks, it works pretty well.