Placeholder system for rendering an action block (#65)
Setup action blocks that represent the place and action inputs.

By place and action inputs, I am referring to the ability to place blocks in the world and remove them. The highlight and placement block will act as visual aids for debugging.

For now this just tracks a block that is in front of the camera by ~8 spaces. This works great for confirming the visual coordinate system is working as intended, but the next steps will be confirming that the highlighted block matches the block coordinate in the logical system.

Previously there were some issues with this, and I am not completely certain they have been resolved.