Support rudimentary block detection at a given position (#66)
Early implementation of block detection within a chunk.

The focus of this is just getting a block at a given location, rather than acting on it. Later work will involve codifying the idea of a detection ray for determining the action block.

This detection will lay the foundations of cleaning up the 'Chunk' exposure in the world, and instead focus on a world-coordinate approach for manipulating coordinates on the fly. The internals of the world such as generation will still rely on chunks, as they are a good way of representing segments of the world.