Place highlight and placement blocks to action blocks (#68)
Place the highlight and placement blocks based on the action positions.

Launch a ray into the scene to attempt to determine the next non-air (e.g. visible) block that an action can be performed on. This could be split into an 'IsDestroyable' or 'CanPlace' checks, but for now Visible is sufficient.