Support block destruction and placement (#69)
Support block destruction and placement with a fixed type.

The block for placement is fixed as dirt, and will need a UI element for block rotations. This way we can place multiple types and blocks and have a visual indicator for this.

This requires an immediate rebuild of the chunk mesh when the state is changed. This is less than ideal, as it doesn't allow batching of alterations to a scene. The concept of a 'dirty state' fits this model a bit better, as it allows for scheduling regenerations based on the dirty state of chunks.