Combine the Biome and WorldGenerator concepts to reduce complexity (#71)
Reduce WorldGenerator into the Biome model to remove an unnecessary
layer of complexity in the generation pipeline.

Splitting biome and world generator added more complexity to the
generation model than was necessary. Rather than attempting to pass the
required data for a generator through the biome in a generic manner, it
made more sense to just merge them.

This makes it easier to define strict inputs for the generator, and for
the concept of 'Biome' to be based on the inputs provided to this
generator.