Resolve null-value errors with BlockType in editor (#74)
Validate the state of a BlockType to avoid issues with texture resolution.

The API for texture element has been slightly adjusted to remove the reliance on resolving the texture pack element. This now just returns the TextureID based on the voxel face.

Additionally some cleanup has been performed on the BlockType to remove unused types.