Use JSON file for TexturePack instead of inline (#75)
Use a JSON definition for the texture pack instead of inline definition.

The output from the spritesheet packer is a JSON file that declares the texture positions. Rather than require inlining through a script, this loads the locations from the JSON on the fly.

Later work will need to be done to make the TexturePack pre-compute the UV and dimension vectors.