Configure test runner for Unity project (#78)
Configure running tests when building on the linux environment.

Add a series of tests for the MathHelper to ensure that the conversion between physical and logical space is working as intended.

The tests do not run on all environments due to the increased build times. WebGL was selected as the only target platform as it is the primary deployment target (blockycraft.jrbeverly.dev/play). If tests begin to make use of execution (UnityTest) then this may change.