Consolidate TexturePack into asset directory (#81)
Consolidate the resources for TexturePacks into a single resource.

The Voxel material is tightly coupled to the TexturePack itself. Rather than be a generic 'Voxel' material, it is closely related to the default texture pack. Any alternate TexturePack would likely possess a different material, rather than use a material that is not used for 'generic' use.

This doesn't discredit the idea of a generic use material for a Voxel, just that the I don't expect the work in Unity to ever see one.