Simple means for transitioning between biomes (#83)
A simple model (that has some visual problems) for transitioning between biomes instead of random.

Early work for setting up a transition model for moving between biomes based on conditions. The condition is a simple probability at the moment, but the alterations provide the starter for thinking about transitions between biomes.

The API for this will need to be refactored a fair bit, as the current Chunk / Biome / World Position APIs are starting to show issues with where data is represented.