Migrate Geometry namespace over to the new Engine format (#84)
Migrate Geometry, which is a foundation namespace, over to the new Blockycraft.Engine.

The project in its early stages didn't leverage any form of real project structure. As a result I have done a fair bit of refactoring to try to fix that up. Now that the project is approaching the ends of where I want it to be, I've decided to do some final mass restructuring of the code.

This should make it easier to understand if I ever decide to look at it again.