BlockType Selector using preview images for each BlockType (#88)
Use a BlockType selector controlled by mouse wheel for placing blocks.

Generate a preview image for all of the block types that can be placed by the player. The mouse wheel can be used to scroll up and down the list. The list is provided as-is, not sorted in any order.

The API options for rendering a mesh on the UI yielded some 'odd' results. Rather than spend significant time trying to figure out the various deprecated / undocumented APIs, I just went with pre-rendering a image that will be drawn on the overlay. Ideally I'd render the mesh on the fly on the interface, and I may switch to that now that this is working.