Epsilon adjustment to avoid whitespaces in render (#90)
Increase the size of each individual block to be larger than 1 to avoid whitespace clipping.

Currently from a distance there is a whitespace that appears between the blocks. This is due to how floating point numbers work in game engines like Unity (and in general). Since the calculations are floating, there are times when the blocks will not overlap exactly.

By padding each of the blocks, we ensure that at least within a certain range the blocks will not have this effect.