Generate chunks in the world based on distance to player (#91)
Generate chunks based on proximity to player to ensure there is always a visible element.

When the chunk factory is looking for the next work item, it will now search for the closest chunk to the player. This is done with the WhereAmI lookup that finds the chunk the player is currently in.

The algorithm for finding the nearest point is the brute force implementation. A more efficient tree structure would be needed (such as a BkTree) to improve the performance. As the Space3D interface evolves, I expect at some point I will make the decision to implement a space partition data structure better suited for the API workload.