Simplify the Player movement code, improving control (#92)
Restructure main camera & player to have better movement controls for the scene.

The old movement controls had issues with navigation when trying to explore the scene. The movement direction was based on where the camera was looking. This meant that something like a 'fly-over' wasn't really possible.

This new change allows a better range of motion when flying over the biomes.